Extract abstract JSON import/export test case
